CI 対応: views fixture 再生成・oversized 対応・成人向けゲート契約テスト分割
- views wire fixture / viewsContract.ts に content_labels を反映
- 取得ゲート契約テストを tests/media_adult_gating.rs へ分割(新規1000行超の回避)
- oversized-baseline を更新(runtimeApi.ts +11行: 表示設定コマンド2件、community_node.rs +1行: content_labels)
- clippy collapsible_if 修正、createPost 呼び出しアサーションに content_labels 引数を追加
