+++ apps/desktop/src/lib/api/__fixtures__/views/bookmarked_post_view.json
@@ -17,6 +17,7 @@
     "content": "minimal",
     "content_status": "Missing",
     "attachments": [],
+    "content_labels": [],
     "created_at": 1700000001,
     "reply_to": null,
     "reply_preview": null,
