+++ apps/desktop/src/lib/api/__fixtures__/views/post_view.minimal.json
@@ -15,6 +15,7 @@
   "content": "minimal",
   "content_status": "Missing",
   "attachments": [],
+  "content_labels": [],
   "created_at": 1700000001,
   "reply_to": null,
   "reply_preview": null,
